debugging

+++ src/main/java/tech/danieljones/loansapiexample/persistance/repository/LoanDetailsRepo.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import tech.danieljones.loansapiexample.persistance.model.LoanDetails;
 
 
-public interface LoanDetailsRepo extends JpaRepository<LoanDetails, Long> {
+public interface LoanDetailsRepo extends JpaRepository<LoanDetails, Integer> {
 
 
 }
